@@ -43,13 +43,18 @@
                 <version>3.8.1</version>
                 <inherited>true</inherited>
                 <configuration>
-                    <source>11</source>
-                    <target>11</target>
-                    <processorPath>
+                    <!--<source>17</source>
+                    <target>17</target>-->
+                    <release>17</release>
+                    <compilerArgs>
+                      <arg>--add-opens</arg> 
+                      <arg>java.base/java.lang=ALL-UNNAMED</arg>
+                    </compilerArgs>
+                    <!--<processorPath>
                       <path>
                         <module>com.sun.tools.javac.processing</module> 
                       </path>
-                    </processorPath>
+                    </processorPath>-->
                 </configuration>
             </plugin>
              <plugin>
@@ -259,6 +264,13 @@
             <version>${openjpa.version}</version>
         </dependency>
 
+        <!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-data-jpa -->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-data-jpa</artifactId>
+            <version>3.2.1</version>
+        </dependency>
+
         <dependency>
             <groupId>com.mchange</groupId>
             <artifactId>c3p0</artifactId>
@@ -316,7 +328,7 @@
         <dependency>
           <groupId>org.projectlombok</groupId>
           <artifactId>lombok</artifactId>
-          <version>1.16.20</version>
+          <version>1.18.30</version>
           <scope>compile</scope>
         </dependency>
 
